pyproject.toml
[tool.poetry]
name = "example-aes-256-py"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
marshmallow = "^3.2.0"
pycryptodome = "^3.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
